requirements.txt
mediapipe
opencv-python
numpy
python-dotenv # for reading .env file
flask
pillow